refactor & new methods & bump

+++ lib/mongo_uploader/mount.rb
@@ -14,41 +14,46 @@ module MongoUploader
 
         after_destroy "delete_#{column.to_s}".to_sym
 
+        define_method("#{column.to_s}_id") do
+          with_value(column) { |val| val.split("/").first rescue val }
+        end
+
         define_method("#{column.to_s}=") do |_file|
           obj_id = self.class.mongo_storage.store(_file)
           write_attribute(column.to_sym, "#{obj_id}/#{_file.original_filename}")
         end
 
         define_method("#{column.to_s}") do
-          tmp = read_attribute(column.to_sym)
-          id = tmp.split("/").first rescue tmp
+          return unless id = send("#{column.to_s}_id")
           self.class.mongo_storage.retrieve(id)
         end
 
+        define_method("delete_#{column.to_s}") do
+          return unless id = send("#{column.to_s}_id")
+          self.class.mongo_storage.delete(id)
+        end
+
         define_method("#{column.to_s}_url") do
-          id = read_attribute(column.to_sym)
-          "/mongo/#{id}"
+           with_value(column) { |val| "/mongo/#{val}" }
         end
 
         define_method("#{column.to_s}_name") do
-          url = read_attribute(column.to_sym)
-          url.split("/").last rescue url
+          with_value(column) { |val| val.split("/").last rescue val }
         end
 
-        define_method("#{column.to_s}_id") do
-          url = read_attribute(column.to_sym)
-          url.split("/").first rescue url
-        end
-
-        define_method("delete_#{column.to_s}") do
-          id = send("#{column.to_s}_id")
-          self.class.mongo_storage.delete(id)
+        define_method("#{column.to_s}_present?") do
+          read_attribute(column.to_sym).present?
         end
 
       end
 
       def mongo_storage
         @storage ||= MongoUploader::Storage.new()
+      end
+
+      def with_value(column, &block)
+        return unless value = read_attribute(column.to_sym)
+        yield value
       end
 
     end
